gravitic-celestial/requirements-docker.txt
# Core framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
rank-bm25>=0.2.2
streamlit>=1.34.0
python-dotenv>=1.0.1
pydantic>=1.10,<3
google-genai>=0.3.0
edgartools>=3.0.0
hishel==0.0.30
requests>=2.31.0
beautifulsoup4>=4.12.0

# Production adapters
psycopg2-binary>=2.9.9
redis>=5.0.0
rq>=1.16.0
sentence-transformers>=2.6.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
